Fixed bug in updateSpotStats and omission in reviewFISHClassification
updateSpotStats estimates how many spots there are.  But it added in _all_
manually classified spots to the automatic ones instead of just
manually classified good spots.  This was fixed.
In addition, in allDone button callback, this updating was not saved with
guidata()
I had some file corruption problems with saved version of spotStats after the
doneButton_Callback.  Not sure what this is.