All interval estimates are now funneled through makeSpotCountInterval.m
makeSpotCountInterval now can handle both spotStats{i} inputs and
trainingSet.RF inputs.  It also now outputs a SpotNumEstimate (which
is ignored in trainRFClassifier).
SpotNumEstimate is also calculated elsewhere (in classifySpots) but
it was needed to be makeSpotCountInterval because of the call from
reviewFISHClassification.  One solution for this small redundancy
would be to remove the calculation from classifySpots, but this is minor