Changes to remove masking from SegStack
In the display and also for edge maxima, cutting the image off
sharply at the mask edge is a problem. We don't want any spots
out there, but do want to allow that in the images.
These files have been modified to do that.
Also some changes to continue propagating the NestedDirStructure
(in the training set functions here)