Color changes in reviewFISHClassifiction.m
Modified the blue and yellow shades and also made it so that
spots that are added as bad or made bad have a yellow slash or X
instead of a blue one which marks good spots.
Also put a green boundary to mark the edge of an object in the right hand pane.