Rewrite of spotStatsDataAligning to have a struct as the data structure
The matrix dta structure of spotStatsDataAligning relies on a certain number of dyes
and remembering which dye is in which column.  This is brittle and inflexible.
This new version uses structs which have matrices in them and are linked via
a unique identifier for every specimen in the dataset (e.g. a worm).
A similar function except also allowing for DAPI to be included was placed
in the Auxiliary function repository for lab use.