Changes to redoing training and to reviewing GUI
The Aro_parameters changes reading introduced a bug in trainRFClassifier
where when the reviewingGUI called it it was not supposed to rerun
variable selection, but it was.  This is fixed.
Also modified the reviewingGUI - the laplace flitered spot image in the lower left
now shows up.  And the surface plots are now smaller and centered around the spot.